NE-2123: Add catalog generation for v1.3.0 build in Konflux
Bumps the existing catalog to v1.3.0 and adds the new stable-v1.3 channel.
Adds Containerfile.aws-load-balancer-operator-catalog to be used in
Konflux for a new catalog build.
Adds catalog-template from which the file based catalog is generated.
Adds new `make generate-catalog` to be used to generate a new catalog
for Konflux-built operator.