NE-2123: Add catalog generation for v1.3.0 build in Konflux
Adds Containerfile.aws-load-balancer-operator-catalog to be used in
Konflux for a new catalog build.
Adds catalog-template from which the file based catalog is generated.
Adds new `make generate-catalog` to be used to generate a new catalog
for Konflux-built operator.
Updates opm used in the Makefile to v1.52.0

Steps to generate new catalog:
- add entry into the catalog/catalog-template.yaml
- run "make generate-catalog"